requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==5.1.0
